Support multiple message combination
Support the combination of multiple response message content, including SearchQuery, LoaderMessage and Disengaged.
The disengaged message is quite necessary as it marks the end of the conversation and informs the user to recreate it.
The other two types of messages may be just for fun, and if not needed, a message type filter option can be added later.